Final Project/api/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
django-cors-headers==4.0.0
kombu==5.5.3
openpyxl==3.1.2
prompt_toolkit==3.0.51
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0

